requirements.txt
psutil
qobuz-dl
tidal-dl
python-telegram-bot==13.4.1
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
python-dotenv>=0.10
tenacity>=6.0.0
python-magic
psycopg2-binary
lxml
humanize
telegraph
heroku3